requirements.txt
Flask==2.3.2
flask-cors==3.0.10
python-docx==0.8.11
Pillow==10.3.0